+++ dynts/media/dynts/decorator.js
@@ -71,7 +71,7 @@
 			}
     		
             var poptions = {
-                    colors: ["#205497","#2D8633","#B84000","#d18b2c"],
+                    //colors: ["#205497","#2D8633","#B84000","#d18b2c"],
                     grid: {hoverable: true, clickable: true, color: '#00264D', tickColor: '#A3A3A3'},
                     selection: {mode: 'xy', color: '#3399FF'},
                     lines: {show: true, lineWidth: 3},