+++ dynts/web/plugins.py
@@ -27,9 +27,15 @@ __ http://packages.python.org/djpcms/'''
         css = {
             'all': ('dynts/ecoplot/ecoplot.css', skin)
         }
+        #js = ['dynts/flot/excanvas.min.js',
+        #      'dynts/flot/jquery.flot.min.js',
+        #      'dynts/flot/jquery.flot.selection.min.js',
+        #      'dynts/jquery.flot.text.js',
+        #      'dynts/ecoplot/ecoplot.js',
+        #      'dynts/decorator.js']
         js = ['dynts/flot/excanvas.min.js',
-              'dynts/flot/jquery.flot.min.js',
-              'dynts/flot/jquery.flot.selection.min.js',
+              'dynts/flot/jquery.flot.js',
+              'dynts/flot/jquery.flot.selection.js',
               'dynts/jquery.flot.text.js',
               'dynts/ecoplot/ecoplot.js',
               'dynts/decorator.js']